refactor: rename canvas-controls (right sidebar) to tool_controls as we already have a canvas_controls
