+++ apps/flites/lib/screens/overview.dart
@@ -1,7 +1,7 @@
 import 'package:flites/constants/app_sizes.dart';
 import 'package:flites/widgets/blocking_widget/blocking_container.dart';
-import 'package:flites/widgets/canvas_controls/canvas_controls.dart';
-import 'package:flites/widgets/canvas_controls/zoom_controls.dart';
+import 'package:flites/widgets/tool_controls/tool_controls.dart';
+import 'package:flites/widgets/tool_controls/zoom_controls.dart';
 import 'package:flites/widgets/image_editor/image_editor.dart';
 import 'package:flites/widgets/project_file_list/project_file_list_vertical.dart';
 import 'package:flutter/material.dart';
@@ -45,7 +45,7 @@ class _OverviewState extends State<Overview> {
                   ],
                 ),
               ),
-              CanvasControls()
+              ToolControls()
             ],
           ),
           BlockingContainer(),
