+++ apps/flites/lib/widgets/tool_controls/zoom_controls.dart
@@ -2,6 +2,7 @@ import 'package:flites/constants/app_sizes.dart';
 import 'package:flites/main.dart';
 import 'package:flites/states/canvas_controller.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
 
 class ZoomControls extends StatelessWidget {
   const ZoomControls({super.key});
@@ -24,7 +25,7 @@ class ZoomControls extends StatelessWidget {
             icon: Icon(
               Icons.zoom_out,
               size: Sizes.p24,
-              color: context.colors.surfaceContainer,
+              color: context.colors.onSurface,
             ),
           ),
           IconButton(
@@ -35,7 +36,7 @@ class ZoomControls extends StatelessWidget {
             icon: Icon(
               Icons.zoom_in,
               size: Sizes.p24,
-              color: context.colors.surfaceContainer,
+              color: context.colors.onSurface,
             ),
           ),
         ],
